Bug fix.
Server wasn't handling clients with improperly formed test requests.  (i.e., telnet'ing to test port would cause
server to kill itself).  The fix was to catch the return code from the initialize_tests() routine.  Now this 
routine returns a negative number on failure and a positive number on success.  The return code is then 
checked in the web100srv.c file and if negative, the child is killed and the server loops back to see if another
client has arrived.

Incremented to ver 3.5.8

Rich