Convert wait() to waidpid() function in testoptions.c file.  This call is made after the c2s & s2c tests run, to catch/close the pkt-pair child process.  The wait() call responded to any child, while the waitpid() call only responds to a specific child.  This may fix a bug with multi-client mode where the server gets multiple signals.
remove a possible extraneous call to child_sig() when a client is listed as stuck in the fifo queue.
The mlab servers are entering a state where a new client is delayed from entering the run state if
a previous client pushed the parent into this stuck state.  

Update version in configure.ac and .java files to 3.5.11

RAC 10/9/09